% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_survival_cpp_M1}
\alias{get_survival_cpp_M1}
\alias{get_survival_cpp_M2}
\alias{get_survival_cpp_M3}
\alias{simulate_M1}
\alias{simulate_M2}
\alias{simulate_M3}
\title{Disease Trajectories and Patient Survival}
\usage{
get_survival_cpp_M1(
  RHO = 5,
  RAISE_KILLING = 1,
  LOWER_GROWTH = 1,
  DIAGNOSIS_THRESHOLD = 6.5e+09,
  DEATH_THRESHOLD = 1e+12,
  TREATMENT_DELAY = 0,
  TREATMENT_DURATION = 1e+10,
  CHEMO_DELAY = 0,
  CHEMO_DURATION = 1e+10
)

get_survival_cpp_M2(
  RHO = 0.04495,
  RAISE_KILLING = 1,
  LOWER_GROWTH = 1,
  DIAGNOSIS_THRESHOLD = 6.5e+09,
  DEATH_THRESHOLD = 1e+12,
  TREATMENT_DELAY = 0,
  TREATMENT_DURATION = 1e+10,
  CHEMO_DELAY = 0,
  CHEMO_DURATION = 1e+10
)

get_survival_cpp_M3(
  RHO = 0.04495,
  RAISE_KILLING = 1,
  LOWER_GROWTH = 1,
  DIAGNOSIS_THRESHOLD = 6.5e+09,
  DEATH_THRESHOLD = 1e+12,
  TREATMENT_DELAY = 0,
  TREATMENT_DURATION = 1e+10,
  CHEMO_DELAY = 0,
  CHEMO_DURATION = 1e+10
)

simulate_M1(
  T_MAX = 120,
  RAISE_KILLING = 1,
  LOWER_GROWTH = 1,
  RHO = 5,
  DIAGNOSIS_THRESHOLD = 6.5e+09,
  DEATH_THRESHOLD = 1e+12,
  TREATMENT_DELAY = 0,
  TREATMENT_DURATION = 1e+10,
  CHEMO_DELAY = 0,
  CHEMO_DURATION = 1e+10
)

simulate_M2(
  T_MAX = 120,
  RAISE_KILLING = 1,
  LOWER_GROWTH = 1,
  RHO = 0.04495,
  DIAGNOSIS_THRESHOLD = 6.5e+09,
  DEATH_THRESHOLD = 1e+12,
  TREATMENT_DELAY = 0,
  TREATMENT_DURATION = 1e+10,
  CHEMO_DELAY = 0,
  CHEMO_DURATION = 1e+10
)

simulate_M3(
  T_MAX = 120,
  RAISE_KILLING = 1,
  LOWER_GROWTH = 1,
  RHO = 0.04495,
  DIAGNOSIS_THRESHOLD = 6.5e+09,
  DEATH_THRESHOLD = 1e+12,
  TREATMENT_DELAY = 0,
  TREATMENT_DURATION = 1e+10,
  CHEMO_DELAY = 0,
  CHEMO_DURATION = 1e+10
)
}
\arguments{
\item{RHO}{Tumor proliferation rate}

\item{RAISE_KILLING}{The multiplication factor to increase the killing rate 
of cytotoxic T cells as a treatment effect of immune therapy. The default 
value 1 indicates no treatment effect. To simulate the immune therapy effect
the value should be greater than 1.}

\item{LOWER_GROWTH}{The multiplication factor to reduce 
the tumor growth as a treatment effect of chemotherapy. The default 
value 1 indicates no treatment effect. To simulate the chemotherapy effect
the value should be lower than 1.}

\item{DIAGNOSIS_THRESHOLD}{The number of tumor cell at which cancer is diagnosed}

\item{DEATH_THRESHOLD}{The lethal number of tumor cell}

\item{TREATMENT_DELAY}{The delay between the diagnosis and start of 
the immune therapy in months}

\item{TREATMENT_DURATION}{Immune therapy duration in months}

\item{CHEMO_DELAY}{The delay between the diagnosis and start of 
the chemotherapy in months}

\item{CHEMO_DURATION}{Chemotherapy duration in months}

\item{T_MAX}{The maximum treatment duration in months}
}
\value{
For the get_survival_cpp_MX functions, the
   survival of a patient in months. For the simulate_MX functions, a
   matrix representing the disease trajectory (see examples). 
   Caution: for the simulate_MX functions the 
output time resolution is _days_, not months.
}
\description{
The functions get_survival_cpp_MX calculate survival of a simulated patient in months
with model MX (X=1,2, or 3), whereas the functions simulate_MX 
generate simulated disease trajectories (i.e., patients over time) 
for a single patient with different models. 
See Creemers et al. (2023) for full definitions of each model.
}
\examples{
# Model M1
# Get survival when no treatment is implemented using the default model parameters
get_survival_cpp_M1()

# Get survival when partial response to immunotherapy is achieved 
# using the default model parameters
get_survival_cpp_M1(RAISE_KILLING=10)

# Get survival when complete response to immunotherapy is achieved  
# using the default model parameters
get_survival_cpp_M1(RAISE_KILLING=20)

# Get survival when partial response to chemotherapy is achieved 
# using the default model parameters
get_survival_cpp_M1(LOWER_GROWTH=0.8)

# Get survival when complete response to chemotherapy is achieved  
# using the default model parameters
get_survival_cpp_M1(LOWER_GROWTH=0.19)

# Get survival when chemotherapy is followed by immunotherapy  
# using the default model parameters
get_survival_cpp_M1(LOWER_GROWTH=0.5, CHEMO_DURATION=10, RAISE_KILLING=15, 
   TREATMENT_DELAY=10, TREATMENT_DURATION=30)

# Model M2
# Get survival when no treatment is implemented using the default model parameters
get_survival_cpp_M2()

# Get survival when partial response to immunotherapy is achieved 
# using the default model parameters
get_survival_cpp_M2(RAISE_KILLING=14000)

# Get survival when complete response to immunotherapy is achieved  
# using the default model parameters
get_survival_cpp_M2(RAISE_KILLING=14500)

# Get survival when partial response to chemotherapy is achieved 
# using the default model parameters
get_survival_cpp_M2(LOWER_GROWTH=0.5)

# Get survival when complete response to chemotherapy is achieved 
# using the default model parameters
get_survival_cpp_M2(LOWER_GROWTH=0.01)

# Model M3
# Get survival when no treatment is implemented using the default model parameters
get_survival_cpp_M3()

# Get survival when partial response to immunotherapy is achieved 
# using the default model parameters
get_survival_cpp_M3(RAISE_KILLING=10)

# Get survival when complete response to immunotherapy is achieved  
# using the default model parameters
get_survival_cpp_M3(RAISE_KILLING=17)

# Get survival when partial response to chemotherapy is achieved 
# using the default model parameters
get_survival_cpp_M3(LOWER_GROWTH=0.8)

# Get survival when complete response to chemotherapy is achieved 
# using the default model parameters
get_survival_cpp_M3(LOWER_GROWTH=0.3)
# Simulates the disease trajectory when no treatment is implemented
t_max <- 12 * 5 # five years
y <- simulate_M1(t_max, 1, 1)
# Transform output to extract time, tumor size and intratumoral T cells
y <- matrix(y, ncol=3, byrow=TRUE) 
# Visualize the disease course in a period shortly before treatment and until death
tumor_size_start <- 1e8
t_start <- which.max(y[,2] >= 1e8) - 1
y <- y[-(1:t_start),]
tumor_size_death <- 1e12
y <- y[y[,2] <= tumor_size_death,] 
day_in_month <- 30.4
y[,1] <- (y[,1] - min(y[,1])) / 30.4
# Plot tumor size over time 
plot( y[,1], y[,2]/1e6, log="y", type="l", col="orange", 
    xlab="time (months)", ylab="tumor burden (million cells)")
}
\references{
Jeroen Creemers, Ankur Ankan, Kit C.B. Roes, Gijs Schröder, Niven Mehra, Carl G. Figdor, 
I. Jolanda M. de Vries, and Johannes Textor (2013), In silico cancer immunotherapy trials
uncover the consequences of therapy-specific response patterns for clinical trial design
and outcome, medRxiv 2021.09.09.21263319; doi: https://doi.org/10.1101/2021.09.09.21263319
}
